// src/ADempiere/modules/window/WindowType.d.ts
export enum EntityEventType {
    INSERT = 0,
    UPDATE = 1,
    DELETE = 2
}
// ...
export enum WorkflowStateType {
    RUNNING = 0,
    COMPLETED = 1,
    ABORTED = 2,
    TERMINATED = 3,
    SUSPENDED = 4,
    NOT_STARTED = 5
}
export enum PriorityType {
    URGENT = 0,
    HIGH = 1,
    MEDIUM = 2,
    LOW = 3,
    MINOR = 4
}

export enum EventType {
    PROCESS_CREATED = 0,
    PROCESS_COMPLETED = 1,
    STATE_CHANGED = 2
}

export interface IEntityLogData {
    logId: number
    id: number
    uuid: string
    tableName: string
    sessionUuid: string
    userUuid: string
    userName: string
    transactionName: string
    eventType: EntityEventType
    eventTypeName: string
    logDate: number
    changeLogsList: IEntityLogData
}
// ...
export interface IWorkflowEventData {
    nodeUuid: string
    nodeName: string
    recordId: number
    tableName: string
    userUuid: string
    userName: string
    responsibleUuid: string
    responsibleName: string
    textMessage: string
    timeElapsed: number
    attributeName: string
    oldValue: string
    newValue: string
    workflowState: WorkflowStateType
    workflowStateName: string
    eventType: EventType
    eventTypeName: string
    logDate: number
}

export interface IWorkflowProcessData {
    processUuid: string
    workflowUuid: string
    workflowName: string
    recordId: number
    tableName: string
    userUuid: string
    userName: string
    responsibleUuid: string
    responsibleName: string
    textMessage: string
    processed: boolean
    workflowStateName: string
    workflowState: WorkflowStateType
    priority: PriorityType
    priorityName: string
    workflowEventsList: IWorkflowEventData[]
    logDate: number
}